Relocate BuilderInput to ord_neu.rs
