+++ differential-dataflow/src/trace/implementations/mod.rs
@@ -56,8 +56,7 @@ pub use self::ord_neu::VecOrdKeyBuilder as KeyBuilder;
 use std::convert::TryInto;
 
 use serde::{Deserialize, Serialize};
-use timely::container::{DrainContainer, PushInto};
-use timely::progress::Timestamp;
+use timely::container::PushInto;
 
 use crate::lattice::Lattice;
 use crate::difference::Semigroup;
@@ -286,82 +285,6 @@ impl BatchContainer for OffsetList {
 
     fn len(&self) -> usize {
         self.len()
-    }
-}
-
-/// Behavior to split an update into principal components.
-pub trait BuilderInput<K: BatchContainer, V: BatchContainer>: DrainContainer + Sized {
-    /// Key portion
-    type Key<'a>: Ord;
-    /// Value portion
-    type Val<'a>: Ord;
-    /// Time
-    type Time;
-    /// Diff
-    type Diff;
-
-    /// Split an item into separate parts.
-    fn into_parts<'a>(item: Self::Item<'a>) -> (Self::Key<'a>, Self::Val<'a>, Self::Time, Self::Diff);
-
-    /// Test that the key equals a key in the layout's key container.
-    fn key_eq(this: &Self::Key<'_>, other: K::ReadItem<'_>) -> bool;
-
-    /// Test that the value equals a key in the layout's value container.
-    fn val_eq(this: &Self::Val<'_>, other: V::ReadItem<'_>) -> bool;
-
-    /// Count the number of distinct keys, (key, val) pairs, and total updates.
-    fn key_val_upd_counts(chain: &[Self]) -> (usize, usize, usize);
-}
-
-impl<K,KBC,V,VBC,T,R> BuilderInput<KBC, VBC> for Vec<((K, V), T, R)>
-where
-    K: Ord + Clone + 'static,
-    KBC: for<'a> BatchContainer<ReadItem<'a>: PartialEq<&'a K>>,
-    V: Ord + Clone + 'static,
-    VBC: for<'a> BatchContainer<ReadItem<'a>: PartialEq<&'a V>>,
-    T: Timestamp + Lattice + 'static,
-    R: Ord + Semigroup + 'static,
-{
-    type Key<'a> = K;
-    type Val<'a> = V;
-    type Time = T;
-    type Diff = R;
-
-    fn into_parts<'a>(((key, val), time, diff): Self::Item<'a>) -> (Self::Key<'a>, Self::Val<'a>, Self::Time, Self::Diff) {
-        (key, val, time, diff)
-    }
-
-    fn key_eq(this: &K, other: KBC::ReadItem<'_>) -> bool {
-        KBC::reborrow(other) == this
-    }
-
-    fn val_eq(this: &V, other: VBC::ReadItem<'_>) -> bool {
-        VBC::reborrow(other) == this
-    }
-
-    fn key_val_upd_counts(chain: &[Self]) -> (usize, usize, usize) {
-        let mut keys = 0;
-        let mut vals = 0;
-        let mut upds = 0;
-        let mut prev_keyval = None;
-        for link in chain.iter() {
-            for ((key, val), _, _) in link.iter() {
-                if let Some((p_key, p_val)) = prev_keyval {
-                    if p_key != key {
-                        keys += 1;
-                        vals += 1;
-                    } else if p_val != val {
-                        vals += 1;
-                    }
-                } else {
-                    keys += 1;
-                    vals += 1;
-                }
-                upds += 1;
-                prev_keyval = Some((key, val));
-            }
-        }
-        (keys, vals, upds)
     }
 }
